Split _stats/_performance.py by concern; drop the % from the layout reason
Addresses #906 and #905.

#905 — the `reason` string in [tool.check_test_layout] contained a literal
`100%`, and tools that read [tool.*] tables through configparser treat `%`
as interpolation syntax and crash. `uvx radon cc src` and `radon mi` both
raised ValueError from the repo root; the /rhiza:quality complexity metrics
for this repo had to be gathered from a different working directory.
Reworded to "the full line-and-branch pytest coverage gate", with a comment
so the character does not come back. radon now runs from the repo root.

#906 — split _stats/_performance.py (714 lines, the largest module in the
repo) along two of the four section markers its author had already written:

  _concentration.py  _ConcentrationStatsMixin  hhi_positive, hhi_negative
  _benchmark.py      _BenchmarkStatsMixin      r_squared, information_ratio,
                                               greeks, treynor_ratio

Both extracted sections were self-contained — neither calls a sibling method
on the mixin — and _BenchmarkStatsMixin has a genuine shared precondition
(every metric needs self._data.benchmark and raises NoBenchmarkError without
it), which is what makes it a unit rather than an arbitrary slice.

Both new mixins are composed into Stats alongside the original, so the public
API is unchanged and no test needed touching: 1231 passed, 100% line+branch
coverage on all three modules. `_probabilistic_ratio_from_base` deliberately
stays in _performance.py, where two tests reference it directly.

Note the issue's stated target — no module over ~400 lines — is NOT met.
_performance.py lands at 468. The remaining ratio family is kept together on
purpose: the probabilistic_*, smart_* and adjusted_* methods are all defined
in terms of the base sharpe and sortino ratios above them, so splitting
further would separate derived metrics from what they derive from.
_performance.py is no longer the largest module; _reporting.py (589) and
_plots/_portfolio.py (588) now are, and both are untouched here.

Also removed two imports (NoBenchmarkError, _comp_return) that the extraction
left unused in _performance.py.
